Add a dio_test project and move request integration test
* run tests for all adapters
* fix redirect bug for `Http2Adapter`
