+++ dio_tests/lib/src/utils.dart
@@ -1,0 +1,3 @@
+const isWeb = bool.hasEnvironment('dart.library.js_util')
+    ? bool.fromEnvironment('dart.library.js_util')
+    : identical(0, 0.0);
