+++ dio_tests/test/adapter/web.dart
@@ -1,0 +1,7 @@
+import 'package:dio/browser.dart';
+import 'package:dio/dio.dart';
+import 'package:dio/io.dart';
+
+final List<HttpClientAdapter> adapters = [
+  BrowserHttpClientAdapter(),
+];
